fix Apotheosis Integration move Apotheosis button injection to separate class to prevent Servers from loading Client only classes
