+++ gradle.properties
@@ -9,7 +9,7 @@ parchmentVersion=2022.05.02
 # mods.toml data
 minecraftVersionCheck=[1.18.2,1.19)
 # Mod Information see https://mcforge.readthedocs.io/en/1.18.x/gettingstarted/versioning/#examples
-modVersion=1.0.0.5
+modVersion=1.0.0.8
 modId=unordinary_basics
 
 jei_version=9.7.0.194
