+++ src/main/java/com/github/manasmods/unordinary_basics/integration/apotheosis/ApotheosisIntegration.java
@@ -1,8 +1,6 @@
-package com.github.manasmods.unordinary_basics.integration;
+package com.github.manasmods.unordinary_basics.integration.apotheosis;
 
-import com.github.manasmods.manascore.client.gui.widget.ImagePredicateButton;
 import com.github.manasmods.unordinary_basics.Unordinary_Basics;
-import com.github.manasmods.unordinary_basics.core.AbstractContainerScreenAccessor;
 import com.github.manasmods.unordinary_basics.core.FletchingContainerAccessor;
 import com.github.manasmods.unordinary_basics.menu.FletchingTableMenu;
 import com.github.manasmods.unordinary_basics.network.Unordinary_BasicsNetwork;
@@ -14,14 +12,12 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.SimpleMenuProvider;
 import net.minecraft.world.inventory.ContainerLevelAccess;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.network.NetworkHooks;
 import shadows.apotheosis.village.fletching.ApothFletchingBlock;
 import shadows.apotheosis.village.fletching.FletchingContainer;
-import shadows.apotheosis.village.fletching.FletchingScreen;
 
 public class ApotheosisIntegration {
-    private static final ResourceLocation UNORDINARY_BASICS_LOGO = new ResourceLocation(Unordinary_Basics.MOD_ID, "textures/gui/ub_logo.png");
+    static final ResourceLocation UNORDINARY_BASICS_LOGO = new ResourceLocation(Unordinary_Basics.MOD_ID, "textures/gui/ub_logo.png");
 
     public ApotheosisIntegration() {
         Unordinary_Basics.getLogger().info("Apotheosis integration automatically enabled!");
@@ -58,15 +54,5 @@ public class ApotheosisIntegration {
             buffer.writeBlockPos(pos);
             buffer.writeResourceLocation(player.level.dimension().location());
         });
-    }
-
-    public void onOpenApotheosisMenu(final ScreenEvent.InitScreenEvent.Post e) {
-        if (!(e.getScreen() instanceof FletchingScreen screen)) return;
-        int imageWidth = ((AbstractContainerScreenAccessor) screen).getImageWidth();
-        e.addListener(new ImagePredicateButton(screen.getGuiLeft() + imageWidth - 24 - 4, screen.getGuiTop() + 6, 24, 24, UNORDINARY_BASICS_LOGO, pButton -> {
-            Unordinary_Basics.getInstance().getApotheosisIntegration().ifPresent(ApotheosisIntegration::requestUnordinaryBasicsFletchingMenu);
-        }, (pButton, pPoseStack, pMouseX, pMouseY) -> {
-            screen.renderTooltip(pPoseStack, new TranslatableComponent(Unordinary_Basics.MOD_ID + ".menu.fletching_table.switch.ub"), pMouseX, pMouseY);
-        }, () -> Unordinary_Basics.getInstance().getApotheosisIntegration().isPresent()));
     }
 }
